Rename files missed previously (to comply with refactoring done earliear to LMS folder)
