+++ src/language-server/task-list/semantic/task-list-identity-reconciler.ts
@@ -4,7 +4,7 @@ import { AstRootNode } from '../../../langium-model-server/semantic/model'
 import type { Initialized } from '../../../langium-model-server/workspace/documents'
 import type * as ast from '../../generated/ast'
 import type * as source from '../lms/model'
-import type { TaskListModelUpdateCalculators } from '../lms/task-list-source-update-calculation'
+import type { TaskListModelUpdateCalculators } from '../lms/task-list-model-update-calculation'
 import type { TaskListServices } from '../task-list-module'
 import { type TaskListDocument } from '../workspace/documents'
 import { Model } from './task-list-identity'
