+++ src/language-server/task-list/task-list-module.ts
@@ -8,8 +8,8 @@ import { TaskListIdentityStorage } from './semantic/task-list-identity-storage'
 import { TaskListSemanticDomain } from './semantic/task-list-semantic-domain'
 import type * as source from './lms/model'
 import { TaskListLangiumModelServerFacade } from './lms/task-list-facade'
-import { TaskListModelUpdateCombiner } from './lms/task-list-source-update-combiner'
-import { TaskListModelUpdateCalculators } from './lms/task-list-source-update-calculation'
+import { TaskListModelUpdateCombiner } from './lms/task-list-model-update-combiner'
+import { TaskListModelUpdateCalculators } from './lms/task-list-model-update-calculation'
 import type { TaskListDocument } from './workspace/documents'
 import { isTaskListDocument } from './workspace/documents'
